requirement.txt
django==3.0.5
djangorestframework==3.12.2
django-cors-headers==3.7.0
djangorestframework-simplejwt
Pillow